refactor(installer): extract update marker region verbatim
